COM-1813 - 2 screens to handle save
